context restructured
