+++ src/frontend/context/LikedContext.js
@@ -1,136 +1,66 @@
-import axios from 'axios';
 import { createContext, useContext, useReducer, useEffect } from 'react';
-import { GETLIKED } from '../routes/routes';
 import { useAuthCtx } from './AuthenticationContext';
+import { addLikedVideo, deleteLikedVideo, getLikedVideos } from '../service';
+import { likedReducerFunc, defaultLikedState } from '../helpers';
 
 const LikedContext = createContext();
-
-const defaultLikedState = {
-  likedLoader: false,
-  likedList: [],
-  addedVideosId: []
-};
-
-const likedReducerFunc = (state, action) => {
-  switch (action.type) {
-    case 'LIKE_API_REQUEST':
-      return {
-        ...state,
-        likedLoader: true
-      };
-    case 'LIKE_API_RESPONSE':
-      return {
-        ...state,
-        likedList: [...action.payload],
-        likedLoader: false
-      };
-    case 'UPDATE_ID':
-      return {
-        ...state,
-        addedVideosId: [...action.payload]
-      };
-    default:
-      return {
-        ...state
-      };
-  }
-};
 
 const LikedProvider = ({ children }) => {
   const [state, dispatch] = useReducer(likedReducerFunc, defaultLikedState);
   const { token } = useAuthCtx();
 
-  const getLikedList = async () => {
-    dispatch({ type: 'LIKE_API_REQUEST' });
-    try {
-      const storedLikes = JSON.parse(localStorage.getItem('userData'))?.likes;
-      if (storedLikes) {
-        dispatch({ type: 'LIKE_API_RESPONSE', payload: storedLikes });
-
-        const idArray = storedLikes.map((elem) => elem._id);
-        dispatch({ type: 'UPDATE_ID', payload: [...idArray] });
-      } else {
-        const {
-          data: { likes }
-        } = await axios.get(GETLIKED);
-        dispatch({ type: 'LIKE_API_RESPONSE', payload: [...likes] });
-
-        const idArray = likes.map((elem) => elem._id);
-        dispatch({ type: 'UPDATE_ID', payload: [...idArray] });
-
-        const datatoUpdate = JSON.parse(localStorage.getItem('userData'));
-        datatoUpdate.likes = [...likes];
-        localStorage.setItem('userData', JSON.stringify(datatoUpdate));
-      }
-    } catch (err) {
-      console.log('LIKED_GET_REQUEST_ERROR', err);
-    }
-  };
-
-  const deleteLikedVideo = async (id, video) => {
+  const deleteLikedFromList = async (id) => {
     dispatch({ type: 'LIKE_API_REQUEST' });
     const { addedVideosId } = state;
-    try {
-      const {
-        data: { likes }
-      } = await axios.delete(`${GETLIKED}/${id}`, {
-        headers: {
-          authorization: token
-        }
-      });
-      const datatoUpdate = JSON.parse(localStorage.getItem('userData'));
-      datatoUpdate.likes = [...likes];
-      localStorage.setItem('userData', JSON.stringify(datatoUpdate));
-      dispatch({ type: 'LIKE_API_RESPONSE', payload: [...likes] });
-      dispatch({
-        type: 'UPDATE_ID',
-        payload: addedVideosId.filter((e) => e !== id)
-      });
-    } catch (err) {
-      console.log('LIKED_DELETE_REQUEST_ERROR', err);
-    }
+
+    const likes = await deleteLikedVideo(id, token);
+    const datatoUpdate = JSON.parse(localStorage.getItem('userData'));
+    datatoUpdate.likes = [...likes];
+    localStorage.setItem('userData', JSON.stringify(datatoUpdate));
+    dispatch({ type: 'LIKE_API_RESPONSE', payload: [...likes] });
+    dispatch({
+      type: 'UPDATE_ID',
+      payload: addedVideosId.filter((e) => e !== id)
+    });
   };
 
   const addToLikedlist = async (video) => {
     dispatch({ type: 'LIKE_API_REQUEST' });
-
     const { addedVideosId } = state;
 
     if (addedVideosId.includes(video._id)) {
-      deleteLikedVideo(video._id, video);
+      deleteLikedFromList(video._id);
     } else {
-      try {
-        const {
-          data: { likes }
-        } = await axios.post(
-          GETLIKED,
-          { video },
-          {
-            headers: {
-              authorization: token
-            }
-          }
-        );
-        const datatoUpdate = JSON.parse(localStorage.getItem('userData'));
-        datatoUpdate.likes = [...likes];
-        localStorage.setItem('userData', JSON.stringify(datatoUpdate));
-        dispatch({ type: 'LIKE_API_RESPONSE', payload: [...likes] });
-        
-        const idArray = likes.map((elem) => elem._id);
-        dispatch({ type: 'UPDATE_ID', payload: [...idArray] });
-      } catch (err) {
-        console.log('LIKED_POST_REQUEST_ERROR', err);
-      }
+      const likes = await addLikedVideo(video, token);
+      const datatoUpdate = JSON.parse(localStorage.getItem('userData'));
+      datatoUpdate.likes = [...likes];
+      localStorage.setItem('userData', JSON.stringify(datatoUpdate));
+      dispatch({ type: 'LIKE_API_RESPONSE', payload: [...likes] });
+      const idArray = likes.map((elem) => elem._id);
+      dispatch({ type: 'UPDATE_ID', payload: [...idArray] });
     }
   };
 
   useEffect(() => {
-    getLikedList();
-  }, []);
+    const getLikedList = async () => {
+      dispatch({ type: 'LIKE_API_REQUEST' });
+
+      const likes = await getLikedVideos(token);
+      dispatch({ type: 'LIKE_API_RESPONSE', payload: [...likes] });
+
+      const idArray = likes.map((elem) => elem._id);
+      dispatch({ type: 'UPDATE_ID', payload: [...idArray] });
+
+      const datatoUpdate = JSON.parse(localStorage.getItem('userData'));
+      datatoUpdate.likes = [...likes];
+      localStorage.setItem('userData', JSON.stringify(datatoUpdate));
+    };
+    if (token) getLikedList();
+  }, [token]);
 
   return (
     <LikedContext.Provider
-      value={{ state, dispatch, addToLikedlist, deleteLikedVideo }}
+      value={{ state, dispatch, addToLikedlist, deleteLikedFromList }}
     >
       {children}
     </LikedContext.Provider>
